remove unsigned

+++ src/iter/iter5.c
@@ -41,7 +41,7 @@ static void altmin_nlop(iter_op_data* _o, int N, float* args[N], unsigned long d
 {
 	const auto data = CAST_DOWN(iter5_altmin_s, _o);
 
-	assert((unsigned int)N == operator_nr_args(data->nlop->op));
+	assert(N == operator_nr_args(data->nlop->op));
 
 	nlop_generic_apply_select_derivative_unchecked(data->nlop, N, (void*)args, der_out, der_in);
 }
