+++ src/nlops/chain.c
@@ -375,8 +375,8 @@ struct nlop_s* nlop_combine(const struct nlop_s* a, const struct nlop_s* b)
 
 	auto cop = operator_combi_create(2, (const struct operator_s*[]){ a->op, b->op });
 
-	assert(II == (int)operator_nr_in_args(cop));
-	assert(OO == (int)operator_nr_out_args(cop));
+	assert(II == operator_nr_in_args(cop));
+	assert(OO == operator_nr_out_args(cop));
 
 	int perm[II + OO];	// ao ai bo bi -> ao bo ai bi
 	int p = 0;
